Tests handle missing environment variables without issue, specifically nil S3_BUCKETs
