+++ spec/controllers/api/v1/icons_controller_spec.rb
@@ -1,8 +1,10 @@
 require "spec_helper"
+require "support/s3_bucket_helper"
 
 RSpec.describe Api::V1::IconsController do
   describe "POST s3_delete", show_in_doc: true do
     it "should require login" do
+      handle_s3_bucket
       expect(S3_BUCKET).not_to receive(:delete_objects)
       post :s3_delete
       expect(response).to have_http_status(401)
@@ -10,6 +12,7 @@ RSpec.describe Api::V1::IconsController do
     end
 
     it "should require s3_key param" do
+      handle_s3_bucket
       expect(S3_BUCKET).not_to receive(:delete_objects)
       login
       post :s3_delete
@@ -18,6 +21,7 @@ RSpec.describe Api::V1::IconsController do
     end
 
     it "should require your own icon" do
+      handle_s3_bucket
       user = create(:user)
       login_as(user)
 
@@ -29,6 +33,7 @@ RSpec.describe Api::V1::IconsController do
     end
 
     it "should not allow deleting a URL in use" do
+      handle_s3_bucket
       icon = create(:uploaded_icon)
       login_as(icon.user)
       expect(S3_BUCKET).not_to receive(:delete_objects)
@@ -38,6 +43,7 @@ RSpec.describe Api::V1::IconsController do
     end
 
     it "should delete the URL" do
+      handle_s3_bucket
       icon = build(:uploaded_icon)
       login_as(icon.user)
       delete_key = {delete: {objects: [{key: icon.s3_key}], quiet: true}}
